pruebabackendjr/requirements.txt
Django
djangorestframework
django-cors-headers
pandas
xlrd